requirements.txt
burnysc2==5.0.6
